requirements.txt
aioboto3==10.3.0 ; python_full_version == "3.10.12"
aiobotocore[boto3]==2.4.1 ; python_full_version == "3.10.12"
aiohttp==3.8.5 ; python_full_version == "3.10.12"
aioitertools==0.11.0 ; python_full_version == "3.10.12"
aiosignal==1.3.1 ; python_full_version == "3.10.12"
annotated-types==0.5.0 ; python_full_version == "3.10.12"
async-timeout==4.0.3 ; python_full_version == "3.10.12"
attrs==23.1.0 ; python_full_version == "3.10.12"
boto3==1.24.59 ; python_full_version == "3.10.12"
botocore==1.27.59 ; python_full_version == "3.10.12"
brotlipy==0.7.0 ; python_full_version == "3.10.12"
cachetools==5.3.1 ; python_full_version == "3.10.12"
certifi==2023.7.22 ; python_full_version == "3.10.12"
cffi==1.15.1 ; python_full_version == "3.10.12"
charset-normalizer==3.2.0 ; python_full_version == "3.10.12"
contourpy==1.1.0 ; python_full_version == "3.10.12"
cryptography==41.0.3 ; python_full_version == "3.10.12"
cycler==0.11.0 ; python_full_version == "3.10.12"
filelock==3.12.3 ; python_full_version == "3.10.12"
fonttools==4.42.1 ; python_full_version == "3.10.12"
frozenlist==1.4.0 ; python_full_version == "3.10.12"
gmpy2==2.1.5 ; python_full_version == "3.10.12"
google-api-core==2.11.1 ; python_full_version == "3.10.12"
google-api-core[grpc]==2.11.1 ; python_full_version == "3.10.12"
google-auth==2.22.0 ; python_full_version == "3.10.12"
google-cloud-bigquery==3.11.4 ; python_full_version == "3.10.12"
google-cloud-core==2.3.3 ; python_full_version == "3.10.12"
google-crc32c==1.5.0 ; python_full_version == "3.10.12"
google-resumable-media==2.6.0 ; python_full_version == "3.10.12"
googleapis-common-protos==1.60.0 ; python_full_version == "3.10.12"
grpcio-status==1.58.0 ; python_full_version == "3.10.12"
grpcio==1.58.0 ; python_full_version == "3.10.12"
idna==3.4 ; python_full_version == "3.10.12"
jinja2==3.1.2 ; python_full_version == "3.10.12"
jmespath==1.0.1 ; python_full_version == "3.10.12"
kiwisolver==1.4.5 ; python_full_version == "3.10.12"
markupsafe==2.1.3 ; python_full_version == "3.10.12"
matplotlib==3.7.2 ; python_full_version == "3.10.12"
mpmath==1.3.0 ; python_full_version == "3.10.12"
multidict==6.0.4 ; python_full_version == "3.10.12"
networkx==3.1 ; python_full_version == "3.10.12"
numpy==1.25.2 ; python_full_version == "3.10.12"
packaging==23.1 ; python_full_version == "3.10.12"
pandas==2.1.0 ; python_full_version == "3.10.12"
pillow==10.0.0 ; python_full_version == "3.10.12"
proto-plus==1.22.3 ; python_full_version == "3.10.12"
protobuf==4.24.3 ; python_full_version == "3.10.12"
pyasn1-modules==0.3.0 ; python_full_version == "3.10.12"
pyasn1==0.5.0 ; python_full_version == "3.10.12"
pycparser==2.21 ; python_full_version == "3.10.12"
pydantic-core==2.6.3 ; python_full_version == "3.10.12"
pydantic-settings==2.0.3 ; python_full_version == "3.10.12"
pydantic==2.3.0 ; python_full_version == "3.10.12"
pyopenssl==23.2.0 ; python_full_version == "3.10.12"
pyparsing==3.0.9 ; python_full_version == "3.10.12"
pysocks==1.7.1 ; python_full_version == "3.10.12"
python-dateutil==2.8.2 ; python_full_version == "3.10.12"
python-dotenv==1.0.0 ; python_full_version == "3.10.12"
pytz==2023.3.post1 ; python_full_version == "3.10.12"
requests==2.31.0 ; python_full_version == "3.10.12"
rsa==4.9 ; python_full_version == "3.10.12"
s3transfer==0.6.2 ; python_full_version == "3.10.12"
six==1.16.0 ; python_full_version == "3.10.12"
sympy==1.12 ; python_full_version == "3.10.12"
torch==2.0.1 ; python_full_version == "3.10.12"
torchvision==0.15.2 ; python_full_version == "3.10.12"
typing-extensions==4.7.1 ; python_full_version == "3.10.12"
tzdata==2023.3 ; python_full_version == "3.10.12"
urllib3==1.26.16 ; python_full_version == "3.10.12"
wrapt==1.15.0 ; python_full_version == "3.10.12"
yarl==1.9.2 ; python_full_version == "3.10.12"
